Implimentado tratamento de erro na class animal e cachorro

+++ projetoFinal-POO/projeto-final-POO/src/main/java/org/example/Animal/Animal.java
@@ -1,4 +1,4 @@
-package Animal;
+package org.example.Animal;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -11,6 +11,13 @@ public abstract class Animal {
     protected String raca;
 
     public Animal(int id, String nome, int idade, String raca) {
+        if (id <= 0) {
+            throw new IllegalArgumentException("ID deve ser positivo.");
+        }
+        if (idade < 0) {
+            throw new IllegalArgumentException("Idade não pode ser negativa.");
+        }
+        
         this.id = id;
         this.nome = nome;
         this.idade = idade;
@@ -38,13 +45,20 @@ public abstract class Animal {
     }
 
     public static void deleteAnimal(Connection conn, int idAnimal) throws SQLException {
+        if (conn == null) {
+            throw new SQLException("Conexão nula.");
+        }
+        
         String sql = "DELETE FROM Animal WHERE idAnimal = ?";
 
         try (PreparedStatement st = conn.prepareStatement(sql)) {
             st.setInt(1, idAnimal);
-            int rows = st.executeUpdate();
+            
+            st.executeUpdate();
 
+        } catch (SQLException entrada) {
+            System.err.println("Erro ao executar exclusão SQL: " + entrada.getMessage());
+            throw entrada; 
         }
     }
 }
-
